+++ source/world/entities/Particles/Particle.h
@@ -130,7 +130,7 @@ __CLASS_NEW_DECLARE(Particle, const ParticleDefinition* particleDefinition, cons
 
 void Particle_constructor(Particle this, const ParticleDefinition* particleDefinition, const SpriteDefinition* spriteDefinition, int lifeSpan, fix19_13 mass);
 void Particle_destructor(Particle this);
-void Particle_update(Particle this, u16 elapsedTime, void (* behavior)(Particle particle));
+void Particle_update(Particle this, int elapsedTime, void (* behavior)(Particle particle));
 void Particle_transform(Particle this);
 void Particle_updateVisualRepresentation(Particle this, bool updateSpritePosition);
 void Particle_addForce(Particle this, const Force* force);
@@ -139,7 +139,7 @@ void Particle_setMass(Particle this, fix19_13 mass);
 void Particle_show(Particle this);
 void Particle_hide(Particle this);
 bool Particle_moves(Particle this);
-u8 Particle_canMoveOverAxis(Particle this, const Acceleration* acceleration);
+int Particle_canMoveOverAxis(Particle this, const Acceleration* acceleration);
 u16 Particle_getHeight(Particle this);
 u16 Particle_getDepth(Particle this);
 Gap Particle_getGap(Particle this);
